nit: Move builtins types to builtins file
