+++ program-test/src/lib.rs
@@ -11,8 +11,9 @@ use {
     solana_banks_server::banks_server::start_local_server,
     solana_program_runtime::InstructionProcessor,
     solana_runtime::{
-        bank::{Bank, Builtin, ExecuteTimings},
+        bank::{Bank, ExecuteTimings},
         bank_forks::BankForks,
+        builtins::Builtin,
         commitment::BlockCommitmentCache,
         genesis_utils::{create_genesis_config_with_leader_ex, GenesisConfigInfo},
     },
