+++ runtime/src/serde_snapshot.rs
@@ -7,8 +7,9 @@ use {
         accounts_index::{AccountSecondaryIndexes, AccountsIndexConfig},
         ancestors::Ancestors,
         append_vec::{AppendVec, StoredMetaWriteVersion},
-        bank::{Bank, BankFieldsToDeserialize, BankRc, Builtins},
+        bank::{Bank, BankFieldsToDeserialize, BankRc},
         blockhash_queue::BlockhashQueue,
+        builtins::Builtins,
         epoch_stakes::EpochStakes,
         hardened_unpack::UnpackedAppendVecMap,
         rent_collector::RentCollector,
