+++ runtime/src/snapshot_utils.rs
@@ -2,7 +2,8 @@ use {
     crate::{
         accounts_db::AccountShrinkThreshold,
         accounts_index::{AccountSecondaryIndexes, AccountsIndexConfig},
-        bank::{Bank, BankSlotDelta, Builtins},
+        bank::{Bank, BankSlotDelta},
+        builtins::Builtins,
         hardened_unpack::{unpack_snapshot, ParallelSelector, UnpackError, UnpackedAppendVecMap},
         serde_snapshot::{
             bank_from_streams, bank_to_stream, SerdeStyle, SnapshotStorage, SnapshotStorages,
